r/virtual_machine_extension: switching to use the Virtual Machine ID
This allows the Virtual Machine Extension to be detected as requiring recreation
when the Virtual Machine needs to be created. Until 2.0 the existing fields,
`virtual_machine_name`, `resource_group_name` and `location` can still be specified;
but the replacement `virtual_machine_id` field will take precedence.

This also updates the tests to use the new `virtual_machine_id` field and but leaves
one example of the old style to confirm this still works.

Finally this also introduces the `force_update_tag` field - which allows for users
to force a re-run of the Virtual Machine Extension even if the configuration itself
hasn't changed.